% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Build}
\alias{Build}
\title{Build Object}
\usage{
Build(id = NULL, sourceProvenance = NULL, results = NULL, status = NULL,
  options = NULL, finishTime = NULL, timeout = NULL, steps = NULL,
  source = NULL, createTime = NULL, logUrl = NULL, statusDetail = NULL,
  images = NULL, startTime = NULL, logsBucket = NULL, projectId = NULL)
}
\arguments{
\item{id}{Unique identifier of the build}

\item{sourceProvenance}{A permanent fixed identifier for source}

\item{results}{Results of the build}

\item{status}{Status of the build}

\item{options}{Special options for this build}

\item{finishTime}{Time at which execution of the build was finished}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{steps}{Describes the operations to be performed on the workspace}

\item{source}{Describes where to find the source files to build}

\item{createTime}{Time at which the build was created}

\item{logUrl}{URL to logs for this build in Google Cloud Logging}

\item{statusDetail}{Customer-readable message about the current status}

\item{images}{List of images expected to be built and pushed to Google Container}

\item{startTime}{Time at which execution of the build was started}

\item{logsBucket}{Google Cloud Storage bucket where logs should be written (see}

\item{projectId}{ID of the project}
}
\value{
Build object
}
\description{
Build Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A build resource in the Container Builder API.At a high level, a Build describes where to find source code, how to buildit (for example, the builder image to run on the source), and what tag toapply to the built image when it is pushed to Google Container Registry.
}
\seealso{
Other Build functions: \code{\link{projects.builds.create}}
}

