% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Build.substitutions}
\alias{Build.substitutions}
\title{Build.substitutions Object}
\usage{
Build.substitutions()
}
\value{
Build.substitutions object
}
\description{
Build.substitutions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Substitutions data for Build resource.
}
\seealso{
Other Build functions: \code{\link{Build}},
  \code{\link{projects.builds.create}}
}
