% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildOperationMetadata}
\alias{BuildOperationMetadata}
\title{BuildOperationMetadata Object}
\usage{
BuildOperationMetadata(build = NULL)
}
\arguments{
\item{build}{The build that the operation is tracking}
}
\value{
BuildOperationMetadata object
}
\description{
BuildOperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata for build operations.
}

