% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildOptions}
\alias{BuildOptions}
\title{BuildOptions Object}
\usage{
BuildOptions(sourceProvenanceHash = NULL, requestedVerifyOption = NULL)
}
\arguments{
\item{sourceProvenanceHash}{Requested hash for SourceProvenance}

\item{requestedVerifyOption}{Options for a verifiable build with details uploaded to the Analysis API}
}
\value{
BuildOptions object
}
\description{
BuildOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional arguments to enable specific features of builds.
}

