% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildTrigger}
\alias{BuildTrigger}
\title{BuildTrigger Object}
\usage{
BuildTrigger(BuildTrigger.substitutions = NULL, substitutions = NULL,
  description = NULL, createTime = NULL, disabled = NULL,
  triggerTemplate = NULL, filename = NULL, id = NULL, build = NULL)
}
\arguments{
\item{BuildTrigger.substitutions}{The \link{BuildTrigger.substitutions} object or list of objects}

\item{substitutions}{Substitutions data for Build resource}

\item{description}{Human-readable description of this trigger}

\item{createTime}{Time when the trigger was created}

\item{disabled}{If true, the trigger will never result in a build}

\item{triggerTemplate}{Template describing the types of source changes to trigger a build}

\item{filename}{Path, from the source root, to a file whose contents is used for the}

\item{id}{Unique identifier of the trigger}

\item{build}{Contents of the build template}
}
\value{
BuildTrigger object
}
\description{
BuildTrigger Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for an automated build in response to source repositorychanges.
}
\seealso{
Other BuildTrigger functions: \code{\link{BuildTrigger.substitutions}},
  \code{\link{projects.triggers.create}},
  \code{\link{projects.triggers.patch}}
}
