% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuiltImage}
\alias{BuiltImage}
\title{Google Cloud Container Builder API Objects 
Builds container images in the cloud.}
\usage{
BuiltImage(name = NULL, digest = NULL)
}
\arguments{
\item{name}{Name used to push the container image to Google Container Registry, as}

\item{digest}{Docker Registry 2}
}
\value{
BuiltImage object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:31:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudbuildv1.auto/R/cloudbuild_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
BuiltImage Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
BuiltImage describes an image built by the pipeline.
}
