% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Hash}
\alias{Hash}
\title{Hash Object}
\usage{
Hash(value = NULL, type = NULL)
}
\arguments{
\item{value}{The hash value}

\item{type}{The type of hash that was performed}
}
\value{
Hash object
}
\description{
Hash Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container message for hash values.
}

