% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{ListBuildTriggersResponse}
\alias{ListBuildTriggersResponse}
\title{ListBuildTriggersResponse Object}
\usage{
ListBuildTriggersResponse(triggers = NULL)
}
\arguments{
\item{triggers}{BuildTriggers for the project, sorted by create_time descending}
}
\value{
ListBuildTriggersResponse object
}
\description{
ListBuildTriggersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response containing existing BuildTriggers.
}

