% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{ListOperationsResponse}
\alias{ListOperationsResponse}
\title{ListOperationsResponse Object}
\usage{
ListOperationsResponse(nextPageToken = NULL, operations = NULL)
}
\arguments{
\item{nextPageToken}{The standard List next-page token}

\item{operations}{A list of operations that matches the specified filter in the request}
}
\value{
ListOperationsResponse object
}
\description{
ListOperationsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for Operations.ListOperations.
}
