% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_functions.R
\name{projects.builds.get}
\alias{projects.builds.get}
\title{Returns information about a previously requested build.The Build that is returned includes its status (e.g., success or failure,or in-progress), and timing information.}
\usage{
projects.builds.get(projectId, id)
}
\arguments{
\item{projectId}{ID of the project}

\item{id}{ID of the build}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/container-builder/docs/}{Google Documentation}
}
