% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_functions.R
\name{projects.triggers.list}
\alias{projects.triggers.list}
\title{Lists existing BuildTrigger.This API is experimental.}
\usage{
projects.triggers.list(projectId)
}
\arguments{
\item{projectId}{ID of the project for which to list BuildTriggers}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/container-builder/docs/}{Google Documentation}
}
