% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_functions.R
\name{projects.triggers.patch}
\alias{projects.triggers.patch}
\title{Updates an BuildTrigger by its project ID and trigger ID.This API is experimental.}
\usage{
projects.triggers.patch(BuildTrigger, projectId, triggerId)
}
\arguments{
\item{BuildTrigger}{The \link{BuildTrigger} object to pass to this method}

\item{projectId}{ID of the project that owns the trigger}

\item{triggerId}{ID of the BuildTrigger to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/container-builder/docs/}{Google Documentation}

Other BuildTrigger functions: \code{\link{BuildTrigger.substitutions}},
  \code{\link{BuildTrigger}},
  \code{\link{projects.triggers.create}}
}
