% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{RepoSource}
\alias{RepoSource}
\title{RepoSource Object}
\usage{
RepoSource(tagName = NULL, commitSha = NULL, projectId = NULL,
  repoName = NULL, branchName = NULL)
}
\arguments{
\item{tagName}{Name of the tag to build}

\item{commitSha}{Explicit commit SHA to build}

\item{projectId}{ID of the project that owns the repo}

\item{repoName}{Name of the repo}

\item{branchName}{Name of the branch to build}
}
\value{
RepoSource object
}
\description{
RepoSource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
RepoSource describes the location of the source in a Google Cloud SourceRepository.
}
