% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{SourceProvenance}
\alias{SourceProvenance}
\title{SourceProvenance Object}
\usage{
SourceProvenance(SourceProvenance.fileHashes = NULL,
  resolvedRepoSource = NULL, resolvedStorageSource = NULL,
  fileHashes = NULL)
}
\arguments{
\item{SourceProvenance.fileHashes}{The \link{SourceProvenance.fileHashes} object or list of objects}

\item{resolvedRepoSource}{A copy of the build's source}

\item{resolvedStorageSource}{A copy of the build's source}

\item{fileHashes}{Hash(es) of the build source, which can be used to verify that the original}
}
\value{
SourceProvenance object
}
\description{
SourceProvenance Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Provenance of the source. Ways to find the original source, or verify thatsome source was used for this build.
}
\seealso{
Other SourceProvenance functions: \code{\link{SourceProvenance.fileHashes}}
}
