% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{Build}
\alias{Build}
\title{Build Object}
\usage{
Build(Build.substitutions = NULL, sourceProvenance = NULL,
  createTime = NULL, images = NULL, projectId = NULL, finishTime = NULL,
  logUrl = NULL, options = NULL, source = NULL, statusDetail = NULL,
  status = NULL, timeout = NULL, results = NULL, logsBucket = NULL,
  steps = NULL, buildTriggerId = NULL, id = NULL, substitutions = NULL,
  startTime = NULL)
}
\arguments{
\item{Build.substitutions}{The \link{Build.substitutions} object or list of objects}

\item{sourceProvenance}{A permanent fixed identifier for source}

\item{createTime}{Time at which the request to create the build was received}

\item{images}{A list of images to be pushed upon the successful completion of all build}

\item{projectId}{ID of the project}

\item{finishTime}{Time at which execution of the build was finished}

\item{logUrl}{URL to logs for this build in Google Cloud Logging}

\item{options}{Special options for this build}

\item{source}{Describes where to find the source files to build}

\item{statusDetail}{Customer-readable message about the current status}

\item{status}{Status of the build}

\item{timeout}{Amount of time that this build should be allowed to run, to second}

\item{results}{Results of the build}

\item{logsBucket}{Google Cloud Storage bucket where logs should be written (see}

\item{steps}{Describes the operations to be performed on the workspace}

\item{buildTriggerId}{The ID of the BuildTrigger that triggered this build, if it was}

\item{id}{Unique identifier of the build}

\item{substitutions}{Substitutions data for Build resource}

\item{startTime}{Time at which execution of the build was started}
}
\value{
Build object
}
\description{
Build Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A build resource in the Container Builder API.At a high level, a Build describes where to find source code, how to buildit (for example, the builder image to run on the source), and what tag toapply to the built image when it is pushed to Google Container Registry.Fields can include the following variables which will be expanded when thebuild is created:- $PROJECT_ID: the project ID of the build.- $BUILD_ID: the autogenerated ID of the build.- $REPO_NAME: the source repository name specified by RepoSource.- $BRANCH_NAME: the branch name specified by RepoSource.- $TAG_NAME: the tag name specified by RepoSource.- $REVISION_ID or $COMMIT_SHA: the commit SHA specified by RepoSource or  resolved from the specified branch or tag.
}
\seealso{
Other Build functions: \code{\link{Build.substitutions}},
  \code{\link{projects.builds.create}}
}
