% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildOptions}
\alias{BuildOptions}
\title{BuildOptions Object}
\usage{
BuildOptions(requestedVerifyOption = NULL, sourceProvenanceHash = NULL)
}
\arguments{
\item{requestedVerifyOption}{Requested verifiability options}

\item{sourceProvenanceHash}{Requested hash for SourceProvenance}
}
\value{
BuildOptions object
}
\description{
BuildOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Optional arguments to enable specific features of builds.
}
