% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildStep}
\alias{BuildStep}
\title{BuildStep Object}
\usage{
BuildStep(entrypoint = NULL, id = NULL, dir = NULL, waitFor = NULL,
  env = NULL, args = NULL, name = NULL)
}
\arguments{
\item{entrypoint}{Optional entrypoint to be used instead of the build step image's default}

\item{id}{Optional unique identifier for this build step, used in wait_for to}

\item{dir}{Working directory (relative to project source root) to use when running}

\item{waitFor}{The ID(s) of the step(s) that this build step depends on}

\item{env}{A list of environment variable definitions to be used when running a step}

\item{args}{A list of arguments that will be presented to the step when it is started}

\item{name}{The name of the container image that will run this particular build step}
}
\value{
BuildStep object
}
\description{
BuildStep Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
BuildStep describes a step to perform in the build pipeline.
}
