% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_objects.R
\name{BuildTrigger}
\alias{BuildTrigger}
\title{BuildTrigger Object}
\usage{
BuildTrigger(filename = NULL, triggerTemplate = NULL, build = NULL,
  createTime = NULL, id = NULL)
}
\arguments{
\item{filename}{Path, from the source root, to a file whose contents is used for the}

\item{triggerTemplate}{Template describing the types of source changes to trigger a build}

\item{build}{Contents of the build template}

\item{createTime}{Time when the trigger was created}

\item{id}{Unique identifier of the trigger}
}
\value{
BuildTrigger object
}
\description{
BuildTrigger Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration for an automated build in response to source repositorychanges.
}
\seealso{
Other BuildTrigger functions: \code{\link{projects.triggers.create}},
  \code{\link{projects.triggers.patch}}
}

