% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudbuild_functions.R
\docType{package}
\name{cloudbuild_googleAuthR}
\alias{cloudbuild_googleAuthR}
\alias{cloudbuild_googleAuthR-package}
\title{Google Cloud Container Builder API
Builds container images in the cloud.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:59:21
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudbuildv1.auto/R/cloudbuild_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}

